requirements.txt
amqp==2.5.2
asgiref==3.2.3
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
Django==2.2.9
django-auth-ldap==2.1.0
django-celery-beat==1.6.0
django-celery-results==1.2.0
django-cors-headers==3.2.1
django-pandas==0.6.1
django-redis==4.11.0
django-timezone-field==3.1
django-webpack-loader==0.6.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-tracking==1.5.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
kombu==4.6.7
mysqlclient==1.4.6
numpy==1.18.1
pandas==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
python-crontab==2.4.0
python-dateutil==2.8.1
python-ldap==3.2.0
pytz==2019.3
redis==3.4.1
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
vine==1.3.0
XlsxWriter==1.2.7
zipp==2.2.0
